python_books/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
cors==1.0.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
drf-flex-fields==1.0.2
filelock==3.13.4
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
idna==3.7
PySocks==1.7.1
requests==2.31.0
requests-file==2.0.0
sqlparse==0.4.4
tldextract==5.1.2
typing_extensions==4.11.0
urllib3==2.2.1
zope.event==5.0
zope.interface==6.2
